Add permission verification

+++ app/api/resources/book.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from ..models.book_info import BookModel
 from ..schema.book_sha import book_args_valid
 from ..utils.format import res
@@ -16,24 +16,32 @@ class BookList(Resource):
 
     @jwt_required()
     def post(self):
-        parser = reqparse.RequestParser()
-        book_args_valid(parser)
-        data = parser.parse_args()
-        try:
-            book_name = data['book_name']
-            author = data['author']
-            text = data['text']
-            image_url = data['image_url']
-            number = data['number']
-            category_id = data['category_id']
-            product_id = data['product_id']
-            book_info = BookModel(book_name=book_name, author=author, text=text, image_url=image_url,
-                                  current_number=number, number=number, category_id=category_id,
-                                  product_id=product_id)
-            book_info.add()
-            return res(message="Add Book successfully!")
-        except Exception as e:
-            return res(success=False, message="Error: {}".format(e), code=500)
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+
+        # 管理员可以执行该操作
+        if role == 'admin':
+            parser = reqparse.RequestParser()
+            book_args_valid(parser)
+            data = parser.parse_args()
+            try:
+                book_name = data['book_name']
+                author = data['author']
+                text = data['text']
+                image_url = data['image_url']
+                number = data['number']
+                category_id = data['category_id']
+                product_id = data['product_id']
+                book_info = BookModel(book_name=book_name, author=author, text=text, image_url=image_url,
+                                      current_number=number, number=number, category_id=category_id,
+                                      product_id=product_id)
+                book_info.add()
+                return res(message="Add Book successfully!")
+            except Exception as e:
+                return res(success=False, message="Error: {}".format(e), code=500)
+
+        else:
+            return res(success=False, message='Access denied.', code=403)
 
 
 class Book(Resource):
@@ -46,32 +54,48 @@ class Book(Resource):
 
     @jwt_required()
     def delete(self, book_id):
-        try:
-            BookModel.delete_by_book_id(book_id)
-            return res(message='Book deleted successfully!')
-        except Exception as e:
-            return res(success=False, message="Error: {}".format(e), code=500)
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+
+        # 管理员可以执行该操作
+        if role == 'admin':
+            try:
+                BookModel.delete_by_book_id(book_id)
+                return res(message='Book deleted successfully!')
+            except Exception as e:
+                return res(success=False, message="Error: {}".format(e), code=500)
+
+        else:
+            return res(success=False, message='Access denied.', code=403)
 
     @jwt_required()
     def put(self, book_id):
-        parser = reqparse.RequestParser()
-        book_args_valid(parser)
-        data = parser.parse_args()
-        try:
-            book_info = BookModel.find_by_book_id(book_id)
-            if book_info:
-                book_name = data['book_name']
-                author = data['author']
-                text = data['text']
-                image_url = data['image_url']
-                current_number = data['current_number']
-                number = data['number']
-                category_id = data['category_id']
-                product_id = data['product_id']
-                book_info.update_book(book_id, book_name, author, text, image_url, current_number, number,
-                                      category_id, product_id)
-                return res(message="Update Book successfully!")
+        jwt_data = get_jwt()
+        role = jwt_data['role']
+
+        # 管理员可以执行该操作
+        if role == 'admin':
+            parser = reqparse.RequestParser()
+            book_args_valid(parser)
+            data = parser.parse_args()
+            try:
+                book_info = BookModel.find_by_book_id(book_id)
+                if book_info:
+                    book_name = data['book_name']
+                    author = data['author']
+                    text = data['text']
+                    image_url = data['image_url']
+                    current_number = data['current_number']
+                    number = data['number']
+                    category_id = data['category_id']
+                    product_id = data['product_id']
+                    book_info.update_book(book_id, book_name, author, text, image_url, current_number, number,
+                                          category_id, product_id)
+                    return res(message="Update Book successfully!")
+                else:
+                    return res(success=False, message="Book not found", code=404)
+            except Exception as e:
+                return res(success=False, message="Error: {}".format(e), code=500)
+
             else:
-                return res(success=False, message="Book not found", code=404)
-        except Exception as e:
-            return res(success=False, message="Error: {}".format(e), code=500)
+                return res(success=False, message='Access denied.', code=403)
